pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ffmpegapi"
version = "0.1.0"
description = "A FastAPI wrapper for FFmpeg"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi>=0.121.1",
    "pydantic-settings>=2.11.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "uvicorn[standard]>=0.23.2",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pre-commit>=4.4.0",
    "pytest>=9.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.hatch.build.targets.wheel]
packages = ["app/"]
